level08.lesson11.bonus01

+++ src/com/javarush/test/level10/lesson11/home01/Solution.java
@@ -10,7 +10,7 @@ public class Solution
     {
         int a = 50;
         int b = 17;
-        double d = a / b;
+        double d = (float)a / b;
         System.out.println(d);
     }
 }
